python_server/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
python-dotenv==1.1.1
httpx==0.28.1
ollama==0.5.1
scikit-learn==1.7.1
python-dateutil==2.9.0.post0
jsonschema==4.25.0
pytz==2025.2

# dev
pytest==8.4.1
pytest-asyncio==0.23.8